requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
django-admin-relation-links==0.2.5
django-js-asset==1.2.2
django-mptt==0.11.0
django-seed==0.2.2
djangorestframework==3.12.2
Faker==5.0.0
kombu==5.0.2
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
vine==5.0.0
wcwidth==0.2.5
